+++ public/index.html
@@ -5,9 +5,7 @@
     <link rel="icon" href="%PUBLIC_URL%/favicon.ico" />
     <meta name="viewport" content="width=device-width, initial-scale=1" />
     <meta name="theme-color" content="#000000" />
-    <meta name="description" content="Kushagra Gupta - Profile Page" />
-    <link rel="apple-touch-icon" href="%PUBLIC_URL%/logo192.png" />
-    <link rel="manifest" href="%PUBLIC_URL%/manifest.json" />
+    <meta name="description" content="Kushagra Gupta - Profile Page" />  
     <title>Profile - Kushagra Gupta</title>
   </head>
   <body>
