+++ src/index.css
@@ -1,9 +1,45 @@
-@tailwind base;
-@tailwind components;
-@tailwind utilities;
+:root {
+  --blue-50: #e1f5fe;
+  --blue-100: #b3e5fc;
+  --blue-200: #81d4fa;
+  --blue-300: #4fc3f7;
+  --blue-400: #29b6f6;
+  --blue-500: #03a9f4;
+  --blue-600: #039be5;
+  --blue-700: #0288d1;
+  --blue-800: #0277bd;
+  --blue-900: #01579b;
+  --blue-A100: #80d8ff;
+  --blue-A200: #40c4ff;
+  --blue-A400: #00b0ff;
+  --blue-A700: #0091ea;
 
-@import url("./colors.css");
-@import url("./theme.css");
+  --green-50: #e8f5e9;
+  --green-100: #c8e6c9;
+  --green-200: #a5d6a7;
+  --green-300: #81c784;
+  --green-400: #66bb6a;
+  --green-500: #4caf50;
+  --green-600: #43a047;
+  --green-700: #388e3c;
+  --green-800: #2e7d32;
+  --green-900: #1b5e20;
+  --green-A100: #b9f6ca;
+  --green-A200: #69f0ae;
+  --green-A400: #00e676;
+  --green-A700: #00c853;
+
+  --gray-50: #fafafa;
+  --gray-100: #f5f5f5;
+  --gray-200: #eeeeee;
+  --gray-300: #e0e0e0;
+  --gray-400: #bdbdbd;
+  --gray-500: #9e9e9e;
+  --gray-600: #757575;
+  --gray-700: #616161;
+  --gray-800: #424242;
+  --gray-900: #212121;
+}
 
 body {
   margin: 0;
@@ -15,8 +51,4 @@ body {
     sans-serif;
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
-  color: #000000;
-  color: var(--font-color);
-  background: #ffffff;
-  background: var(--bg-color);
 }
