+++ src/material/material-icons.js
