--- src/theme.css
@@ -1,21 +1,0 @@
-@import url("./colors.css");
-
-:root {
-  --font-color: var(--blue-900);
-  --link-color: var(--blue-A700);
-  --bg-color: var(--gray-50);
-}
-
-/* light theme */
-[data-theme="light"] {
-  --font-color: var(--blue-900);
-  --link-color: var(--blue-A700);
-  --bg-color: var(--gray-50);
-}
-
-/* dark theme */
-[data-theme="dark"] {
-  --font-color: var(--blue-900);
-  --link-color: var(--blue-A700);
-  --bg-color: var(--gray-900);
-}
